Introduce a working base version of program
Split the program into files for each class,
introduced the basic functionality that I
expected from a working version of this program.
Used ArrayDeque instead of Queue or Stack for
convenience, since it serve as both. From now
on I think I will introduce changes properly,
in form of pull requests. Used Notation and
polymorphism for calling the methods, because
all of the derived classes have the same
public interface, so there's no problem.