Use BigDecimal, fix bugs
This commit will make the program use BigDecimal, instead
of the double primitive. Unfortunately, BigDecimal doesn't
support exponents with decimal power, so I was forced to
use a hack, as I didn't want to add external dependencies.
Should work good enough. Also fixed two bugs: first, where
dot in decimal number was skipped because the line was
valueString += expr[++i] instead of valueString += expr[i++].
Second was the bug in PrefixNotation, which made the expression
evaluate in a wrong order.