Add unary '-'; negative numbers
This commit will add unary '-' to all of the notations.
With infix it works both on numbers and parentheses. However,
do keep in mind that something like 2-2 is taken as '2 - 2' instead
of '2 -2'. In postfix and prefix, '2-2' is taken as '2 -2' always.
Might change that behavior in the future, if it turns out to be
wrong somehow.