Make keybindings local;precision;new binds
Made keybindings local to mainDialog, so they only
are in effect when it has focus. Thus the popup
that appears when something goes wrong can be closed
with a keypress.
Made it so that displayed result is in normal notation,
not engineering.
Added bindings for left/right arrow, which scroll the display.